Add files via upload
The Resume Analyzer & Job Matcher is a smart tool designed to evaluate resumes and match them with the most suitable job profiles. It scans key details like skills, experience, and education to identify job fit. The system uses intelligent parsing and matching algorithms to enhance career opportunities. It's an ideal solution for both students and professionals aiming to find the right job faster and smarter.